brand_generator_backend/requirements.txt
Django==4.1.3
requests
django_rest_framework
djongo
